Configuration changed to AutoConfiguration
