+++ src/main/java/org/qubership/integration/platform/engine/configuration/web/WebAutoConfiguration.java
@@ -17,13 +17,13 @@
 package org.qubership.integration.platform.engine.configuration.web;
 
 import org.qubership.integration.platform.engine.configuration.MDCInterceptor;
-import org.springframework.context.annotation.Configuration;
+import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.PathMatchConfigurer;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
-@Configuration
-public class WebConfig implements WebMvcConfigurer {
+@AutoConfiguration
+public class WebAutoConfiguration implements WebMvcConfigurer {
 
     @Override
     public void addInterceptors(InterceptorRegistry registry) {
