+++ src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,6 +1,9 @@
-org.qubership.integration.platform.engine.configuration.MicroserviceRestTemplateConfiguration
-org.qubership.integration.platform.engine.configuration.camel.CamelSecurityConfiguration
-org.qubership.integration.platform.engine.configuration.datasource.PersistenceCheckpointConfiguration
-org.qubership.integration.platform.engine.configuration.datasource.PersistenceQuartzConfiguration
+org.qubership.integration.platform.engine.configuration.MicroserviceRestTemplateAutoConfiguration
+org.qubership.integration.platform.engine.configuration.camel.CamelSecurityAutoConfiguration
+org.qubership.integration.platform.engine.configuration.datasource.PersistenceCheckpointAutoConfiguration
+org.qubership.integration.platform.engine.configuration.datasource.PersistenceQuartzAutoConfiguration
 org.qubership.integration.platform.engine.configuration.opensearch.OpenSearchConfiguration
-org.qubership.integration.platform.engine.configuration.SwaggerConfiguration
+org.qubership.integration.platform.engine.configuration.SwaggerAutoConfiguration
+org.qubership.integration.platform.engine.configuration.ApplicationAutoConfiguration
+org.qubership.integration.platform.engine.configuration.web.WebAutoConfiguration
+org.qubership.integration.platform.engine.configuration.DeploymentReadinessAutoConfiguration
